SampleRegen on each trial.
