+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/sampling/ConcurrentTaskRunner.kt
@@ -43,7 +43,7 @@ class ConcurrentTaskRunner {
         }
 
         // doesnt return until all tasks are done
-        println("that took ${stopwatch.tookPer(tasks.size, "task")}")
+        println("that ${stopwatch.tookPer(tasks.size, "task")}")
         return results
     }
