+++ core/src/main/kotlin/org/cryptobiotic/rlauxe/sampling/MultiContestTestData.kt
@@ -8,6 +8,101 @@ import kotlin.random.Random
 
 private val debug = false
 
+/*
+    val test = MultiContestTestData(20, 11, 20000)
+    val contests: List<Contest> = test.makeContests()
+    val testCvrs = test.makeCvrsFromContests()
+    val ballots = test.makeBallots()
+
+    val test = MultiContestTestData(20, 11, 20000)
+    val contestsUA: List<ContestUnderAudit> = test.makeContests().map { ContestUnderAudit(it, it.Nc) }
+    val cvrsUAP = test.makeCvrsFromContests().map { CvrUnderAudit.fromCvrIF( it, false) }
+ */
+data class MultiContestTestData(
+    val ncontest: Int, val nballotStyles: Int, val totalBallots: Int,
+    val debug: Boolean = false, val minMargin: Double = 0.005,
+) {
+    val fcontests: List<FuzzedContest>
+    val ballotStyles: List<BallotStyle>
+    var countBallots = 0
+
+    init {
+        // between 2 and 4 candidates, margin random number between minMargin and minMargin + .02
+        fcontests = List(ncontest) { it }.map {
+            val ncands = max(Random.nextInt(5), 2)
+            FuzzedContest(it, ncands, minMargin + Random.nextDouble(0.2))
+        }
+
+        // between 1 and ncontest contests, randomly chosen
+        ballotStyles = List(nballotStyles) { it }.map {
+            val ncInStyle = if (ncontest == 1) 1 else
+                1 + Random.nextInt(ncontest - 1) // TODO use a power law distribution (weighted to small values) ? more realistic ??
+            val contestIndexes = mutableSetOf<Int>()
+            while (contestIndexes.size < ncInStyle) {
+                contestIndexes.add(Random.nextInt(ncontest))
+            }
+            val contestList = contestIndexes.map { fcontests[it].info.name }
+            val contestIds = contestIndexes.map { fcontests[it].info.id }
+            val ncards = (totalBallots.toDouble() / nballotStyles).toInt()
+            countBallots += ncards
+            BallotStyle(contestList, contestIds, ncards)
+        }
+        countCards()
+    }
+
+    fun countCards() {
+        ballotStyles.forEach { bs ->
+            bs.contestNames.forEach { contestName ->
+                val contest = fcontests.find { it.info.name == contestName }!!
+                contest.ncards += bs.ncards
+            }
+        }
+    }
+
+    fun makeContests(): List<Contest> {
+        return fcontests.map { it.makeContest() }
+    }
+
+    fun makeBallots(): List<BallotUnderAudit> {
+        val result = mutableListOf<BallotUnderAudit>()
+        var ballotId = 0
+        ballotStyles.forEach { ballotStyle ->
+            repeat(ballotStyle.ncards) {
+                result.add(BallotUnderAudit(ballotId, ballotStyle))
+                ballotId++
+            }
+        }
+        return result
+    }
+
+    override fun toString() = buildString {
+        appendLine("SampleData(ncontest=$ncontest, nballotStyles=$nballotStyles, totalBallots=$totalBallots")
+        fcontests.forEach { appendLine(it) }
+        ballotStyles.forEach { appendLine(it) }
+    }
+
+    fun makeCvrsFromContests(): List<Cvr> {
+        fcontests.forEach { it.resetTracker() } // startFresh
+        val cvrbs = CvrBuilders().addContests(fcontests.map { it.info })
+        val result = mutableListOf<Cvr>()
+        ballotStyles.forEach { ballotStyle ->
+            val fcontests = fcontests.filter { ballotStyle.contestNames.contains(it.info.name) }
+            repeat(ballotStyle.ncards) {
+                result.add(randomSample(cvrbs, fcontests))
+            }
+        }
+        return result.toList()
+    }
+
+    private fun randomSample(cvrbs: CvrBuilders, fcontests: List<FuzzedContest>): Cvr {
+        val cvrb = cvrbs.addCrv()
+        fcontests.forEach { fcontest ->
+            cvrb.addContest(fcontest.info.name, fcontest.chooseCandidate(Random.nextInt(fcontest.votesLeft))).done()
+        }
+        return cvrb.build()
+    }
+}
+
 data class FuzzedContest(
     val contestId: Int,
     val ncands: Int,
@@ -107,188 +202,3 @@ data class FuzzedContest(
         append(" FuzzedContest(contestId=$contestId, ncands=$ncands, margin=${df(margin)}, choiceFunction=$choiceFunction countCards=$ncards")
     }
 }
-
-/*
-    val test = MultiContestTestData(20, 11, 20000)
-    val contests: List<Contest> = test.makeContests()
-    val testCvrs = test.makeCvrsFromContests()
-    val ballots = test.makeBallots()
-
-    val test = MultiContestTestData(20, 11, 20000)
-    val contestsUA: List<ContestUnderAudit> = test.makeContests().map { ContestUnderAudit(it, it.Nc) }
-    val cvrsUAP = test.makeCvrsFromContests().map { CvrUnderAudit.fromCvrIF( it, false) }
- */
-data class MultiContestTestData(
-    val ncontest: Int, val nballotStyles: Int, val totalBallots: Int,
-    val debug: Boolean = false, val minMargin: Double = 0.005,
-) {
-    val fcontests: List<FuzzedContest>
-    val ballotStyles: List<BallotStyle>
-    var countBallots = 0
-
-    init {
-        // between 2 and 4 candidates, margin random number between minMargin and minMargin + .02
-        fcontests = List(ncontest) { it }.map {
-            val ncands = max(Random.nextInt(5), 2)
-            FuzzedContest(it, ncands, minMargin + Random.nextDouble(0.2))
-        }
-
-        // between 1 and ncontest contests, randomly chosen
-        ballotStyles = List(nballotStyles) { it }.map {
-            val ncInStyle = if (ncontest == 1) 1 else
-                1 + Random.nextInt(ncontest - 1) // TODO use a power law distribution (weighted to small values) ? more realistic ??
-            val contestIndexes = mutableSetOf<Int>()
-            while (contestIndexes.size < ncInStyle) {
-                contestIndexes.add(Random.nextInt(ncontest))
-            }
-            val contestList = contestIndexes.map { fcontests[it].info.name }
-            val contestIds = contestIndexes.map { fcontests[it].info.id }
-            val ncards = (totalBallots.toDouble() / nballotStyles).toInt()
-            countBallots += ncards
-            BallotStyle(contestList, contestIds, ncards)
-        }
-        countCards()
-    }
-
-    fun countCards() {
-        ballotStyles.forEach { bs ->
-            bs.contestNames.forEach { contestName ->
-                val contest = fcontests.find { it.info.name == contestName }!!
-                contest.ncards += bs.ncards
-            }
-        }
-    }
-
-    fun makeContests(): List<Contest> {
-        return fcontests.map { it.makeContest() }
-    }
-
-    fun makeBallots(): List<BallotUnderAudit> {
-        val result = mutableListOf<BallotUnderAudit>()
-        var ballotId = 0
-        ballotStyles.forEach { ballotStyle ->
-            repeat(ballotStyle.ncards) {
-                result.add(BallotUnderAudit(ballotId, ballotStyle))
-                ballotId++
-            }
-        }
-        return result
-    }
-
-    override fun toString() = buildString {
-        appendLine("SampleData(ncontest=$ncontest, nballotStyles=$nballotStyles, totalBallots=$totalBallots")
-        fcontests.forEach { appendLine(it) }
-        ballotStyles.forEach { appendLine(it) }
-    }
-
-    fun makeCvrsFromContests(): List<Cvr> {
-        fcontests.forEach { it.resetTracker() } // startFresh
-        val cvrbs = CvrBuilders().addContests(fcontests.map { it.info })
-        val result = mutableListOf<Cvr>()
-        ballotStyles.forEach { ballotStyle ->
-            val fcontests = fcontests.filter { ballotStyle.contestNames.contains(it.info.name) }
-            repeat(ballotStyle.ncards) {
-                result.add(randomSample(cvrbs, fcontests))
-            }
-        }
-        return result.toList()
-    }
-
-    private fun randomSample(cvrbs: CvrBuilders, fcontests: List<FuzzedContest>): Cvr {
-        val cvrb = cvrbs.addCrv()
-        fcontests.forEach { fcontest ->
-            cvrb.addContest(fcontest.info.name, fcontest.chooseCandidate(Random.nextInt(fcontest.votesLeft))).done()
-        }
-        return cvrb.build()
-    }
-
-    fun makeFuzzedCvrsFrom(contests: List<Contest>, cvrs: List<Cvr>, fuzzPct: Double): List<Cvr> {
-        var count = 0
-        var countf = 0
-        val cvrbs = CvrBuilders.convertCvrs(contests.map { it.info }, cvrs)
-        cvrbs.forEach { cvrb: CvrBuilder ->
-            val r = secureRandom.nextDouble(1.0)
-            cvrb.contests.forEach { (contestId, cvb) ->
-                if (r < fuzzPct) {
-                    countf++
-                    val ccontest: CvrContest = cvb.contest
-                    val currId = cvb.votes[0] // TODO only one vote allowed
-                    cvb.votes.clear()
-
-                    // choose a different candidate, or none.
-                    val ncandId = chooseNewCandidate(currId, ccontest.candidateIds)
-                    if (ncandId != null) {
-                        cvb.votes.add(ncandId)
-                    }
-                }
-            }
-            count++
-        }
-        // println("makeFuzzedCvrsFrom $countf/ $count")
-        return cvrbs.map { it.build() }
-    }
-}
-
-fun chooseNewCandidate(currId: Int, candidateIds: List<Int>): Int? {
-    val size = candidateIds.size
-    while (true) {
-        val ncandIdx = secureRandom.nextInt(size + 1)
-        if (ncandIdx == size) return null
-        val candId = candidateIds[ncandIdx]
-        if (candId != currId) {
-            return candId
-        }
-    }
-}
-
-class ComparisonSamplerRegen(
-    val test: MultiContestTestData,
-    val fuzzPct: Double,
-    val cvrs: List<Cvr>, // (mvr, cvr)
-    val contestUA: ContestUnderAudit,
-    val cassorter: ComparisonAssorter
-): GenSampleFn {
-    val N = cvrs.size
-    val cvrsUA = cvrs.map { CvrUnderAudit(it, false) }
-
-    val permutedIndex = MutableList(N) { it }
-    var cvrPairs: List<Pair<CvrIF, CvrUnderAudit>> // (mvr, cvr)
-    var idx = 0
-
-    init {
-        val mvrs = remakeFuzzed()
-        cvrPairs = mvrs.zip(cvrsUA)
-        //sampleMean = cvrPairs.filter{it.first.hasContest(contestUA.id)}.map { (mvr, cvr) -> cassorter.bassort(mvr, cvr) }.average()
-        //sampleCount = cvrPairs.filter{it.first.hasContest(contestUA.id)}.map { (mvr, cvr) -> cassorter.bassort(mvr, cvr) }.sum() // wtf?
-    }
-
-    override fun sample(): Double {
-        while (idx < cvrPairs.size) {
-            val (mvr, cvr) = cvrPairs[permutedIndex[idx]]
-            if (cvr.hasContest(contestUA.id) && (cvr.sampleNum <= contestUA.sampleThreshold || contestUA.sampleThreshold == 0L)) {
-                val result = cassorter.bassort(mvr, cvr)
-                idx++
-                return result
-            }
-            idx++
-        }
-        throw RuntimeException("no samples left for ${contestUA.id} and ComparisonAssorter ${cassorter}")
-    }
-
-    override fun reset() {
-        val mvrs = remakeFuzzed()
-        cvrPairs = mvrs.zip(cvrsUA)
-        permutedIndex.shuffle(secureRandom)
-        idx = 0
-    }
-
-    fun remakeFuzzed(): List<Cvr> {
-        return test.makeFuzzedCvrsFrom(listOf(contestUA.contest), cvrs, fuzzPct)
-    }
-
-    // TODO where needed ?
-    override fun sampleMean() = 0.0
-    override fun sampleCount() = 0.0
-    override fun N() = N
-}
-
