+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/sampling/PlotFindSampleSizes.kt
@@ -55,47 +55,41 @@ class PlotFindSampleSizes {
     }
 
     @Test
-    fun testPollingFuzz() {
+    fun plotPollingFuzzConcurrent() {
         val N = 10000
-        val test = MultiContestTestData(1, 1, N)
-        val contests: List<Contest> = test.makeContests()
         val fuzzPcts = listOf(0.0, 0.001, .005, .01, .02, .05)
         val margins = listOf(.001, .002, .003, .004, .005, .006, .008, .01, .012, .016, .02, .03, .04, .05, .06, .07, .08, .10)
 
-        val auditConfig = AuditConfig(AuditType.POLLING, riskLimit=0.05, seed = 12356667890L, quantile=.80, ntrials = 100)
+        val auditConfig = AuditConfig(AuditType.POLLING, riskLimit=0.05, seed = 12356667890L, quantile=.80, ntrials = 1000)
         val finder = FindSampleSize(auditConfig)
         println("ntrials = ${auditConfig.ntrials} quantile = ${auditConfig.quantile} N=${N}")
 
-        // need coroutines - too slow
-        val srts = mutableListOf<SRT>()
+        val tasks = mutableListOf<AlphaTask>()
         fuzzPcts.forEach { fuzzPct ->
             margins.forEach { margin ->
-                val stopwatch = Stopwatch()
-                val fcontest = FuzzedContest(0, 2, margin)
+                val fcontest = FuzzedContest(0, 4, margin)
                 fcontest.ncards = N
                 val contest = fcontest.makeContest()
                 val cvrs = fcontest.makeCvrs().map { it as Cvr }
-                val mvrs = test.makeFuzzedCvrsFrom(contests, cvrs, fuzzPct)
 
                 print("fuzzPct = $fuzzPct, margin = $margin ${contest.votes}")
                 val contestUA = ContestUnderAudit(contest, N)
-
-                // polling
                 contestUA.makePollingAssertions()
-                val assort = contestUA.pollingAssertions.first().assorter // TODO only one
-                val sampleFn = PollWithoutReplacement(contestUA, mvrs, assort)
+                val minAssort = contestUA.minPollingAssertion().assorter
+                val sampleFn = PollingSamplerRegen(fuzzPct, cvrs, contestUA, minAssort)
 
                 val otherParameters = mapOf("fuzzPct" to fuzzPct)
-                val result =
-                    finder.simulateSampleSizeAlphaMart(sampleFn, margin, assort.upperBound(), N, N, otherParameters)
-                srts.add(result.makeSRT(N, margin2mean(margin), 0.0))
-
-                println(" took ${stopwatch.elapsed(TimeUnit.MILLISECONDS)} ms")
+                tasks.add( AlphaTask("fuzzPct = $fuzzPct, margin = $margin", finder,
+                    sampleFn, margin, minAssort.upperBound(), N, N, otherParameters,
+                ))
             }
         }
+        // run tasks concurrently
+        val results: List<RunTestRepeatedResult> = ConcurrentTaskRunner().run(tasks)
+        val srts = results.map{ it.makeSRT(N, 0.0, 0.0)}
 
-        val dirName = "/home/stormy/temp/polling"
-        val filename = "PollingFuzzSampleSize"
+        val dirName = "/home/stormy/temp/estimate"
+        val filename = "PollingFuzzConcurrent"
         val writer = SRTcsvWriter("$dirName/${filename}.cvs")
         writer.writeCalculations(srts)
         writer.close()
@@ -105,69 +99,19 @@ class PlotFindSampleSizes {
     }
 
     @Test
-    fun testComparisonFuzz() {
+    fun plotComparisonFuzzConcurrent() {
         val N = 10000
-        val test = MultiContestTestData(1, 1, N)
-        val contests: List<Contest> = test.makeContests()
         val fuzzPcts = listOf(0.0, 0.001, .005, .01, .02, .05)
         val margins = listOf(.001, .002, .003, .004, .005, .006, .008, .01, .012, .016, .02, .03, .04, .05, .06, .07, .08, .10)
 
-        val auditConfig = AuditConfig(AuditType.POLLING, riskLimit=0.05, seed = 12356667890L, quantile=.80, ntrials = 1000)
-        val finder = FindSampleSize(auditConfig)
-        println("ntrials = ${auditConfig.ntrials} quantile = ${auditConfig.quantile} N=${N}")
-
-        val srts = mutableListOf<SRT>()
-        fuzzPcts.forEach { fuzzPct ->
-            margins.forEach { margin ->
-                val stopwatch = Stopwatch()
-                val fcontest = FuzzedContest(0, 2, margin)
-                fcontest.ncards = N
-                val contest = fcontest.makeContest()
-                val cvrs = fcontest.makeCvrs().map { it as Cvr }
-                val mvrs = test.makeFuzzedCvrsFrom(contests, cvrs, fuzzPct)
-
-                print("fuzzPct = $fuzzPct, margin = $margin ${contest.votes}")
-                val contestUA = ContestUnderAudit(contest, N)
-
-                // comparison
-                contestUA.makeComparisonAssertions(cvrs)
-                val cassort = contestUA.comparisonAssertions.first().assorter
-                val cvrsUA = cvrs.map { CvrUnderAudit(it, false) }
-
-                val sampleFn = ComparisonSamplerGen(mvrs.zip(cvrsUA), contestUA, cassort)
-
-                // need coroutines - too slow
-                val otherParameters = mapOf("fuzzPct" to fuzzPct)
-                val result = finder.simulateSampleSizeBetaMart(sampleFn, margin, cassort.noerror, cassort.upperBound(), N, N, otherParameters)
-                srts.add(result.makeSRT(N, margin2mean(margin), 0.0))
-                println(" took ${stopwatch.elapsed(TimeUnit.MILLISECONDS)} ms")
-            }
-        }
-        val dirName = "/home/stormy/temp/polling"
-        val filename = "ComparisonFuzzSampleSize"
-        val writer = SRTcsvWriter("$dirName/${filename}.cvs")
-        writer.writeCalculations(srts)
-        writer.close()
-
-        val plotter = EstimateSampleSize(dirName, filename)
-        plotter.showFuzzedSamples()
-    }
-
-    @Test
-    fun testComparisonFuzzConcurrent() {
-        val N = 10000
-        val test = MultiContestTestData(1, 1, N)
-        val fuzzPcts = listOf(0.0, 0.001, .005, .01, .02, .05)
-        val margins = listOf(.001, .002, .003, .004, .005, .006, .008, .01, .012, .016, .02, .03, .04, .05, .06, .07, .08, .10)
-
-        val auditConfig = AuditConfig(AuditType.POLLING, riskLimit=0.05, seed = 12356667890L, quantile=.80, ntrials = 1000)
+        val auditConfig = AuditConfig(AuditType.CARD_COMPARISON, riskLimit=0.05, seed = 12356667890L, quantile=.80, ntrials = 100)
         val finder = FindSampleSize(auditConfig)
         println("ntrials = ${auditConfig.ntrials} quantile = ${auditConfig.quantile} N=${N}")
 
         val tasks = mutableListOf<BettingTask>()
         fuzzPcts.forEach { fuzzPct ->
             margins.forEach { margin ->
-                val fcontest = FuzzedContest(0, 2, margin)
+                val fcontest = FuzzedContest(0, 4, margin)
                 fcontest.ncards = N
                 val contest = fcontest.makeContest()
                 val cvrs = fcontest.makeCvrs().map { it as Cvr }
@@ -177,12 +121,12 @@ class PlotFindSampleSizes {
 
                 // comparison; regen mvrs each repition to smoothe things out
                 contestUA.makeComparisonAssertions(cvrs)
-                val cassort = contestUA.comparisonAssertions.first().assorter
-                val sampleFn = ComparisonSamplerRegen(test, fuzzPct, cvrs, contestUA, cassort)
+                val minAssort = contestUA.minComparisonAssertion().assorter
+                val sampleFn = ComparisonSamplerRegen(fuzzPct, cvrs, contestUA, minAssort)
 
                 val otherParameters = mapOf("fuzzPct" to fuzzPct)
                 tasks.add( BettingTask("fuzzPct = $fuzzPct, margin = $margin", finder,
-                    sampleFn, margin, cassort.noerror, cassort.upperBound(), N, N, otherParameters,
+                    sampleFn, margin, minAssort.noerror, minAssort.upperBound(), N, N, otherParameters,
                 ))
             }
         }
@@ -190,8 +134,8 @@ class PlotFindSampleSizes {
         val results: List<RunTestRepeatedResult> = ConcurrentTaskRunner().run(tasks)
         val srts = results.map{ it.makeSRT(N, 0.0, 0.0)}
 
-        val dirName = "/home/stormy/temp/polling"
-        val filename = "ComparisonFuzzSampleSizeConcurrent"
+        val dirName = "/home/stormy/temp/estimate"
+        val filename = "ComparisonFuzzConcurrent"
         val writer = SRTcsvWriter("$dirName/${filename}.cvs")
         writer.writeCalculations(srts)
         writer.close()
@@ -213,6 +157,22 @@ class BettingTask(val name: String,
 ): ConcurrentTask {
     override fun name() = name
     override fun run() : RunTestRepeatedResult {
+        //  this uses auditConfig p1,,p4 to set apriori error rates. should be based on fuzzPct i think
         return finder.simulateSampleSizeBetaMart(sampleFn, margin, noerror, upperBound, Nc, Nc, otherParameters)
+    }
+}
+
+class AlphaTask(val name: String,
+                  val finder: FindSampleSize,
+                  val sampleFn: GenSampleFn,
+                  val margin: Double,
+                  val upperBound: Double,
+                  val maxSamples: Int,
+                  val Nc: Int,
+                  val otherParameters: Map<String, Double>,
+): ConcurrentTask {
+    override fun name() = name
+    override fun run() : RunTestRepeatedResult {
+        return finder.simulateSampleSizeAlphaMart(sampleFn, margin, upperBound, Nc, Nc, otherParameters)
     }
 }
